Change previous and next links
The existing `previous` and `next` links seem to point in the wrong
directions.  They are `next` and `previous` in a blog order (i.e.,
chronology), but the sidebar menu presents the links in descending
order.  This could also be fixed by changing the order in the sidebar
menu, but this fix seems more straightforward.