Add content to homepage
The current `index.html` template does not provide any content for the
homepage—it leaves the body of the page blank.  This does not seem to be
consistent with how other book-style themes work; they typically either
display the first section or, more rarely, display custom content.

This change would allow either behaviour: if any content is provided in
`_index.md`, it would display that content.  Otherwise, it would default
to displaying the first section.